requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
aiohttp=3.8.5=py311h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=11.0.0=h374c478_2
async-timeout=4.0.2=py311h06a4308_0
attrs=22.1.0=py311h06a4308_0
aws-c-common=0.6.8=h5eee18b_1
aws-c-event-stream=0.1.6=h6a678d5_6
aws-checksums=0.1.11=h5eee18b_2
aws-sdk-cpp=1.8.185=h721c034_1
blas=1.0=mkl
boost-cpp=1.82.0=hdb19cb5_1
bottleneck=1.3.5=py311hbed6279_0
brotlipy=0.7.0=py311h5eee18b_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=py311h06a4308_0
cffi=1.15.1=py311h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=41.0.2=py311h22a60cf_0
cudatoolkit=11.8.0=h6a678d5_0
cudnn=8.9.2.26=cuda11_0
cupti=11.8.0=he078b1a_0
dataclasses=0.8=pyh6d0b6a4_7
datasets=2.14.5=py_0
dill=0.3.7=py311h06a4308_0
filelock=3.9.0=py311h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py311h5eee18b_0
fsspec=2023.4.0=py311h06a4308_0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h5eee18b_3
glog=0.5.0=h2531618_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
grpc-cpp=1.48.2=he1ff14a_1
huggingface_hub=0.15.1=py311h06a4308_0
icu=73.1=h6a678d5_0
idna=3.4=py311h06a4308_0
importlib-metadata=6.0.0=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py311h06a4308_0
joblib=1.2.0=py311h06a4308_0
jpeg=9e=h5eee18b_1
krb5=1.20.1=h143b758_1
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.82.0=ha8e66a6_1
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcurl=8.2.1=h251f7ec_0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.52.0=h2d74bed_1
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
magma=2.7.1=h2c23e93_0
markupsafe=2.1.1=py311h5eee18b_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.6=py311ha02d727_1
mkl_random=1.2.2=py311ha02d727_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
multidict=6.0.2=py311h5eee18b_0
multiprocess=0.70.15=py311h06a4308_0
ncurses=6.4=h6a678d5_0
networkx=3.1=py311h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numexpr=2.8.4=py311h65dcdc2_1
numpy=1.25.2=py311h08b1b3b_0
numpy-base=1.25.2=py311hf175353_0
openssl=3.0.10=h7f8727e_2
orc=1.7.4=hb3bc3d3_1
packaging=23.1=py311h06a4308_0
pandas=2.0.3=py311ha02d727_0
pillow=9.4.0=py311h6a678d5_0
pip=23.2.1=py311h06a4308_0
pyarrow=11.0.0=py311hd8e8d9b_1
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py311h06a4308_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.4=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py311h5eee18b_1
pytorch=2.0.1=gpu_cuda118py311h7668aad_0
pytz=2022.7=py311h06a4308_0
pyyaml=6.0=py311h5eee18b_1
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2022.7.9=py311h5eee18b_0
requests=2.31.0=py311h06a4308_0
safetensors=0.3.2=py311hb02cf49_0
scikit-learn=1.2.2=py311h6a678d5_1
scipy=1.11.1=py311h08b1b3b_0
setuptools=68.0.0=py311h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
timm=0.9.7=pyhd8ed1ab_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=py311h22610ee_1
torchvision=0.15.2=cuda118py311h4cc2eb7_0
tqdm=4.65.0=py311h92b7b1e_0
transformers=4.32.1=py311h06a4308_0
typing-extensions=4.7.1=py311h06a4308_0
typing_extensions=4.7.1=py311h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wheel=0.38.4=py311h06a4308_0
xxhash=0.8.0=h7f8727e_3
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py311h5eee18b_0
zipp=3.11.0=py311h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
